add an invert button in the algo widget

+++ Code/GUI/lib/TraceEditing/QGoSetOfContoursWaterShedAlgo.h
@@ -129,9 +129,9 @@ protected:
         }
 
       bounds[2*iOrientation]    =
-          iCenter[iOrientation] + (pair*(pow(-1, i))*spacing[iOrientation]);
+          iCenter[iOrientation] + (pair*(pow(-1., static_cast<int>(i) ) )*spacing[iOrientation]);
       bounds[2*iOrientation +1] =
-          iCenter[iOrientation] + (pair*(pow(-1, i))*spacing[iOrientation]);
+          iCenter[iOrientation] + (pair*(pow(-1., static_cast<int>(i) ) )*spacing[iOrientation]);
 
       // then let's extract the Slice of Interest
       ImageType2DPointer ITK_Slice_Image =
